feat: Add labels browser to side bar
